Check and find False to repare

+++ API/Controllers/AdminController.cs
@@ -1,3 +1,12 @@
+using System.Linq;
+using System.Threading.Tasks;
+using API.Entities;
+using API.Interfaces;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+
 namespace API.Controllers
 {
     public class AdminController : BaseApiController
@@ -5,7 +14,8 @@ namespace API.Controllers
         private readonly UserManager<AppUser> _userManager;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IPhotoService _photoService;
-        public AdminController(UserManager<AppUser> userManager, IUnitOfWork unitOfWork, IPhotoService photoService)
+        public AdminController(UserManager<AppUser> userManager, IUnitOfWork unitOfWork, 
+            IPhotoService photoService)
         {
             _photoService = photoService;
             _unitOfWork = unitOfWork;
@@ -31,35 +41,35 @@ namespace API.Controllers
             return Ok(users);
         }
 
-
         [Authorize(Policy = "RequireAdminRole")]
         [HttpPost("edit-roles/{username}")]
-
         public async Task<ActionResult> EditRoles(string username, [FromQuery] string roles)
         {
             var selectedRoles = roles.Split(",").ToArray();
+
             var user = await _userManager.FindByNameAsync(username);
 
             if (user == null) return NotFound("Could not find user");
 
             var userRoles = await _userManager.GetRolesAsync(user);
+
             var result = await _userManager.AddToRolesAsync(user, selectedRoles.Except(userRoles));
+
             if (!result.Succeeded) return BadRequest("Failed to add to roles");
 
             result = await _userManager.RemoveFromRolesAsync(user, userRoles.Except(selectedRoles));
 
-            if (!result.Succeeded) return BadRequest("Failed to remove form roles");
+            if (!result.Succeeded) return BadRequest("Failed to remove from roles");
+
             return Ok(await _userManager.GetRolesAsync(user));
-
         }
-
 
         [Authorize(Policy = "ModeratePhotoRole")]
         [HttpGet("photos-to-moderate")]
-
         public async Task<ActionResult> GetPhotosForModeration()
         {
-            var photos = await _unitOfWork.PhotoRepository.GetUnApprovedPhotos();
+            var photos = await _unitOfWork.PhotoRepository.GetUnapprovedPhotos();
+
             return Ok(photos);
         }
 
@@ -67,10 +77,10 @@ namespace API.Controllers
         [HttpPost("approve-photo/{photoId}")]
         public async Task<ActionResult> ApprovePhoto(int photoId)
         {
-            // APPROVE PHOTO to check to see if the user has any photos that are set to main, if not then set the photo to main when approving.
             var photo = await _unitOfWork.PhotoRepository.GetPhotoById(photoId);
 
             if (photo == null) return NotFound("Could not find photo");
+
             photo.IsApproved = true;
 
             var user = await _unitOfWork.UserRepository.GetUserByPhotoId(photoId);
@@ -78,18 +88,22 @@ namespace API.Controllers
             if (!user.Photos.Any(x => x.IsMain)) photo.IsMain = true;
 
             await _unitOfWork.Complete();
+
             return Ok();
         }
 
         [Authorize(Policy = "ModeratePhotoRole")]
         [HttpPost("reject-photo/{photoId}")]
-
         public async Task<ActionResult> RejectPhoto(int photoId)
         {
             var photo = await _unitOfWork.PhotoRepository.GetPhotoById(photoId);
+
+            if (photo == null) return NotFound("Could not find photo");
+
             if (photo.PublicId != null)
             {
                 var result = await _photoService.DeletePhotoAsync(photo.PublicId);
+
                 if (result.Result == "ok")
                 {
                     _unitOfWork.PhotoRepository.RemovePhoto(photo);
@@ -101,9 +115,9 @@ namespace API.Controllers
             }
 
             await _unitOfWork.Complete();
+
             return Ok();
         }
-
 
     }
 }
